LayoutTests/inspector/debugger/live-edit-expected.txt

Tests live edit feature.

Debugger was enabled.

Running: testLiveEdit
==Source frame contents start==
function f()
{
    return "live-edited string";
}

==Source frame contents end==
Debugger was disabled.
